show comments as icon
